+++ server_backend/app/utility/db_connection.py
@@ -1,0 +1,6 @@
+from flask_mysqldb import MySQL
+
+mysql = MySQL()
+
+def get_db_connection():
+    return mysql.connection
